Backend/textutils/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.10.0
Markdown==3.3.6
MouseInfo==0.1.3
numpy==1.22.0
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.0
pip-review==1.1.1
psycopg2==2.9.3
PyAutoGUI==0.9.53
pycodestyle==2.8.0
PyGetWindow==0.0.9
Pygments==2.11.1
PyMsgBox==1.0.9
pyparsing==3.0.6
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
pytweening==1.0.4
pytz==2021.3
pywhatkit==5.2
requests==2.27.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
urllib3==1.26.7
wikipedia==1.4.0
zipp==3.7.0
